refactor: move ui components
