+++ src/app/proyectos/crear/subida/page.tsx
@@ -1,6 +1,6 @@
 import { DocumentArrowUpIcon } from "@heroicons/react/24/solid"
 
-import { ProyectoFileUploaderForm } from "@/app/ui"
+import { ProyectoFileUploaderForm } from "@/ui"
 
 export default async function Page() {
   return (
