--- src/app/ui/information/index.ts
@@ -1,1 +1,0 @@
-export * from "./proyectoFileUpdate/SubmitStatusInfo"
