+++ src/ui/button/index.ts
@@ -1,3 +1,4 @@
+export * from "./Button"
 export * from "./DeleteButton"
 export * from "./EditButton"
 export * from "./SearchProyectosByInvestigadorButton"
