+++ src/ui/index.ts
@@ -1,8 +1,8 @@
-export * from "./breadcrumbs"
 export * from "./button"
 export * from "./cards"
 export * from "./form"
+export * from "./information"
+export * from "./layout"
 export * from "./menu/Menu"
-export * from "./Pagination"
 export * from "./proyectoByInvestigador/ProyectoViewerByInvestigador"
-export * from "./Search"
+export * from "./search"
